refactor(evaluator): split gym_runtime into a package, address review
Review feedback on #1295, from Nick and from CodeRabbit. The fixes and the
split are one commit because the split relocated the file the fixes changed,
so separating them afterwards would mean reconstructing a file that no longer
exists.

**The all-empty raise ignored unattributed failures.** The guard covered
unattributed *successes* but not failures. Both are invisible to
`len(trials)` for the same reason, neither becomes a trial, and a failure
record is a diagnosis in its own right — an OOM-killed agent, say. A run with
one empty rollout and one unattributed failure raised "No agent ran" despite
the environment having recorded something. Reproduced, fixed, regression test.

**The token-usage keys now have a checked source of truth.** The question was
whether they could come from Gym. They cannot: this runtime never imports
`nemo_gym` (it shells out, and we exclude Ray by constraint), and the names
are not Gym's — they are OpenAI's, from the Chat Completions and Responses
usage schemas, and `openai` is a dependency here. Added a test asserting the
tuple matches both schemas, rather than deriving it at import time: what we
match is wire-format JSON, stable regardless of how that package organises
its modules, and importing a nested path for five string literals would break
the runner if it ever moves.

**`hydra.run.dir` was the one override emitted unquoted.** Every other value
goes through `_hydra_scalar`; this one interpolated a `Path` directly.
Checked against Hydra's own parser: a work dir containing `[` raises
`OverrideParseException`, and — worse — one containing `,` silently parses as
a `ChoiceSweep`, so the run directory would be wrong rather than rejected.
Spaces turn out to be fine, so the risk is narrower than it first looks, but
the value is a path and paths are exactly where those characters appear. Now
quoted at both call sites, with a parametrized regression test.

**A missing results file raised a bare `FileNotFoundError`.** Sibling reads in
that module guard existence or degrade; `_trials_from_rollouts` did not. It is
reachable: `_ensure_fresh_output` deletes any stale file before the run and
`_collect_rollouts` already raised on a non-zero exit, so an absent file means
Gym reported success and wrote nothing. Now names the collection log and the
per-server startup log, like the other diagnostics here.

**The split.** At ~1600 lines the module had outgrown a single file. Now a
package named for the runtime, following the `fabric/` and `codex/` precedent
already in `runtimes/` — `runtime.py` as the entry point, siblings named for
what they own:

  config.py    GymRuntimeConfig and the Hydra grammar it serializes into
  records.py   Gym's on-disk artifacts and the index key that joins them
  dataset.py   Gym rows to tasks, and back to a dataset Gym collects against
  results.py   rollouts read back as trials and scores
  process.py   locating, watching and tearing down the gym CLI
  runtime.py   GymAgentTaskRunner — orchestration only

Largest module is now 543 lines. The relocation itself changes no behaviour:
code was moved by slicing line ranges rather than retyped, so the relocated
line count matches the original exactly.

`records.py` earns its place rather than being a leftovers bin. The
`_ng_task_index` / `_ng_rollout_index` pair is the attribution contract
between the half that *writes* it onto a materialized row (dataset) and the
half that *reads* it back off a rollout (results); one definition is what
stops the writer and reader disagreeing, and putting it in either module
would create a cycle. `config.py` likewise holds the config model together
with its Hydra serialization: splitting them needs a cycle, since
`_selection_args` takes a `GymRuntimeConfig`, and they are two halves of one
idea — what Gym is told, and how it is told.

The import path changes from `runtimes.gym_runtime` to `runtimes.gym`. Only
two non-test call sites existed, both updated; the public names re-export
from `__init__` unchanged. No compatibility shim, on the same reasoning as
the field rename in the follow-up: the runner is beta and young enough that a
permanent alias would cost more than it saves.
